[Docs] Documentation update

+++ include/symtab/functb.h
@@ -24,7 +24,7 @@ typedef struct func_ctx {
 
 int FNTB_get_info_id(long id, func_info_t* out, functab_ctx_t* ctx);
 int FNTB_get_info(string_t* fname, func_info_t* out, functab_ctx_t* ctx);
-int FNTB_add_info(string_t* name, int global, int external, int entry, ast_node_t* args, ast_node_t* rtype, functab_ctx_t* ctx);
+long FNTB_add_info(string_t* name, int global, int external, int entry, ast_node_t* args, ast_node_t* rtype, functab_ctx_t* ctx);
 int FNTB_update_info(long id, int used, int entry, ast_node_t* args, ast_node_t* rtype, functab_ctx_t* ctx);
 int FNTB_unload(functab_ctx_t* ctx);
